requirement.txt
amqp==5.0.9
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
cffi==1.15.0
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
django-mdeditor==0.1.20
django-redis==5.2.0
django-simpleui==2022.3.1
djangorestframework==3.13.1
dnspython==2.2.0
eventlet==0.33.0
greenlet==1.1.2
jwt==1.3.1
kombu==5.2.3
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.7
pytz==2021.3
redis==4.1.4
simpleui==4.0.2
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
